+++ server/controllers/RegisterProfessors.js
@@ -1,6 +1,7 @@
 const ProfessorsModel = require("../models/professorsScheema");
 const SubjectsModel = require("../models/subjectsScheema");
 const fs = require("fs");
+const bcrypt = require("bcrypt");
 
 const RegisterProfessors = async (req, res) => {
   // Function to parse csv file
@@ -17,14 +18,16 @@ const RegisterProfessors = async (req, res) => {
   }
 
   // Function to convert CSV data to objects
-  function convertCSVToObjects(csvData) {
-    return csvData.map((data) => ({
+  async function convertCSVToObjects(csvData) {
+    return await Promise.all(csvData.map(async (data) => ({
       username: data.username,
-      password: data.password,
+      password: await bcrypt.hash(data.password, 10),
       name: data.name,
       address: data.address,
       type: data.type,
-    }));
+      Department: data.Department,
+      Designation: data.Designation,
+    })));
   }
   try {
     if (!req.file) {
@@ -35,7 +38,7 @@ const RegisterProfessors = async (req, res) => {
     const csvData = await parseCSV(req.file.path);
 
     // Convert CSV data to an array of objects suitable for MongoDB
-    const formattedData = convertCSVToObjects(csvData);
+    const formattedData = await convertCSVToObjects(csvData);
 
     // Insert data into MongoDB using the UserModel
     await ProfessorsModel.insertMany(formattedData);
@@ -48,7 +51,9 @@ const RegisterProfessors = async (req, res) => {
     res.status(500).json({ message: "Error uploading CSV data" });
   } finally {
     // Clean up the uploaded file (optional)
-    fs.unlinkSync(req.file.path);
+    if (req && req.file && req.file.path) {
+      fs.unlinkSync(req.file.path);
+    }
   }
 };
